# feat: implement path-compressed trie for IP/CIDR lookups using cidranger