--- a/pkg/dataset/root_test.go
+++ b/pkg/dataset/root_test.go
@@ -1,7 +1,6 @@
 package dataset
 
 import (
-	"net/netip"
 	"testing"
 
 	"github.com/crowdsecurity/crowdsec-spoa/internal/remediation"
@@ -24,15 +23,11 @@ func TestDataSet(t *testing.T) {
 	t.Run("Test Init", func(t *testing.T) {
 		// Test new returns the types we expect
 		assert.NotNil(t, dataSet)
-		assert.IsType(t, &PrefixSet{}, dataSet.PrefixSet)
 		assert.IsType(t, &CNSet{}, dataSet.CNSet)
-		assert.IsType(t, &IPSet{}, dataSet.IPSet)
-		assert.NotNil(t, dataSet.PrefixSet)
+		assert.IsType(t, &CIDRUnifiedIPSet{}, dataSet.CIDRUnifiedIPSet)
 		assert.NotNil(t, dataSet.CNSet)
-		assert.NotNil(t, dataSet.IPSet)
-		assert.IsType(t, map[netip.Prefix]RemediationIdsMap{}, dataSet.PrefixSet.Items)
+		assert.NotNil(t, dataSet.CIDRUnifiedIPSet)
 		assert.IsType(t, map[string]RemediationIdsMap{}, dataSet.CNSet.Items)
-		assert.IsType(t, map[netip.Addr]RemediationIdsMap{}, dataSet.IPSet.Items)
 	})
 	tests := []struct {
 		name     string