Replace map-based IPSet and PrefixSet with unified cidranger-based implementation for significant performance improvements and proper longest-prefix matching.

Performance improvements:
- Lookup: ~915ns/op (small dataset), ~1.3μs/op (large dataset)
- Memory: 523-603 B/op, 9-12 allocations
- Add/Remove: ~17ms/op for 2,000 decisions

Key changes:
- Add cidranger dependency for battle-tested trie implementation
- Replace separate IPSet/PrefixSet with unified CIDRUnifiedIPSet
- Implement proper longest-prefix matching semantics
- Store exact IPs as /32 (IPv4) or /128 (IPv6) prefixes
- Maintain thread safety with RWMutex
- Add comprehensive benchmarks and tests

Files:
- pkg/dataset/cidranger_trie.go: Core trie implementation
- pkg/dataset/cidranger_types.go: Unified IP set wrapper
- pkg/dataset/benchmark_test.go: Performance benchmarks
- pkg/dataset/root.go: Updated to use new implementation
- pkg/dataset/root_test.go: Updated tests

This provides ~18-20x performance improvement over the previous O(n) map-based implementation while maintaining backward compatibility.

**benchmark**

| Implementation | Small Dataset (2K entries) | Large Dataset (20K entries) | Memory Allocations |
|----------------|----------------------------|-----------------------------|-------------------|
| **Current (map-based)** | 16,141 ns/op | 155,162 ns/op | 1000 B/op, 12 allocs |
| **CIDR Library** | 864.4 ns/op | 1,256 ns/op | 522 B/op, 9 allocs |